# backend/doc/api.yaml
openapi: 3.0.0
info:
  title: React Learn Words
  description: a backend part of RSSchool-React "Learn words" task
  version: 1.0.0
servers:
  - url: /
components:
  schemas:
...
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        profileImg:
          type: string
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags:
        - Users
      security:
        []
      summary: Create a new user
      description: Creates a new user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  description: User name
                email:
                  type: string
                  description: E-mail
                password:
                  type: string
                  description: Password
                profileImage:
                  type: string
      responses:
        200:
          description: Successful creation.
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/User"
        422:
          description: Incorrect e-mail or password
  /users/{id}:
    parameters:
      - name: id
        description: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Users
      summary: Get user
      description: Gets user
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/User"
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: User not found
    put:
      tags:
        - Users
      summary: Update a user
      description: Updates a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                email:
                  type: string
                  description: E-mail
                password:
                  type: string
                  description: Password
      responses:
        200:
          description: The user has been updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Bad request
        401:
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Users
      summary: Delete a user
      description: Deletes a user by id
      responses:
        204:
          description: The user has been deleted
        401:
          $ref: '#/components/responses/UnauthorizedError'
...